back/requirements/base.txt
aiofile==3.7.4
alembic==1.8.0
fastapi==0.78.0
fastapi_jwt_auth==0.5.0
packaging==21.3
passlib==1.7.4
pydantic==1.9.1
requests==2.28.1
SQLAlchemy==1.4.39
starlette==0.19.1
starlette_prometheus==0.9.0
python-multipart
uvicorn
asyncpg
psycopg2
psycopg2-binary
python-dateutil